Move updateCreationTimestampOfDirectoryAndChildrenFiles to its own file
